Move data from graph to home component
Move flow from child -> parent to parent -> child, this way the data can be sent from the parent component to the graph aswell as the table
