+++ src/app/Features/overview/components/expense-income-line-chart/expense-income-line-chart.component.html
@@ -1,10 +1,10 @@
-<section *ngIf="(monthlySummary$ | async)?.length > 0" class="financial-overview max-w-3xl">
+<section *ngIf="(data | async)?.length > 0" class="financial-overview max-w-3xl">
       <h2 class="text-2xl font-bold mb-4 text-gray-800 ">Yearly overview</h2>
       <div class="bg-white rounded-lg shadow-lg p-4">
         <ngx-charts-line-chart
             [view]="view"
             [scheme]="colorScheme"
-            [results]="monthlySummary$ | async"
+            [results]="data| async"
             [xAxis]="true"
             [yAxis]="true"
             [legend]="legend"
